requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
django-cors-headers==4.3.1
coverage==7.3.4
pytest
pytest-django
